requirements.txt
django-stdimage
